Extract edit action
